Clean up unused functions and legacy code
Remove deprecated utility functions from core, pipeline, and storage modules that are no longer used. Delete empty test files and unused service module. Refactor report response format from list to dictionary mapping question_id to scores. Update .gitignore to exclude report output directory.

+++ core/scoring_math.py
@@ -22,7 +22,7 @@ def z_to_percentile(z):
     if z is None:
         return None
 
-    return round(norm.cdf(z) * 100, 1)
+    return round(float(norm.cdf(z) * 100), 1)
 
 def interpret_percentile(p):
     if p is None:
